Initial commit
This commit introduces an initial, early version of a CLI, TUI and
client library for the Infralight SaaS. The project is written in Go,
and will be distributed as a single, statically-linked executable (an
OCI image will also be included in the future).

The TUI provides a visually appealing interface to various API
operations. It is not meant to expose all the features provided by the
API (nor the Infralight Dashboard).