Add support for profiles and file-based configs
This commit modifies the CLI to support multiple profiles, each with a
different set of access and secret keys, and potentially API URL and
authorization header. Profiles can be stored in TOML files stored in
the standard $XDG_CONFIG_HOME directory of the user.

The default profile is called "default", and will be used if the
`--profile` flag was not provided. If a configuration file does not
exist, and a key-pair was not provided via command line flags, the user
will be prompted to enter a key-pair as before.

A `configure` command is added to easily create a profile.