Add Drifts view to TUI
This commit adds an implementation for the Drifts tab. When selecting
the tab, a list of drift executions will be loaded and displayed. When a
specific drift is selected (by moving to it with the cursor and clicking
Enter), a list of assets will be loaded and displayed, which each asset
displaying its type, ID and state. States are color-coded, with
unmanaged shown as red, modified as orange, and managed as green.

Also included is a performance improvement: when signing in, the CLI
will save the JWT token for up to 3 minutes in a temporary file,
allowing CI scripts to call the CLI multiple times without having to
authenticate each time.