TUI: add support for the codify API method
This commit extends the TUI to support running Infralight's reverse
learning API (a.k.a codify) on an asset from the drifts tab.

The asset list's viewport module is modified to allow selecting rows
from the viewport. The currently selected row is highlighted with a
purple background. When the user presses Enter, the codify method will
be called for the highlighted asset, and the generated state file will
be displayed.

Note that due to the codify method not working successfully on most
assets, the display code is untested.

Resolves: INFL-304