CLI: add support for the States API
This commit introduces the `states` family of commands to the CLI. Users
can now get the latest state in a stack, or list the states, and even
upload policies.

When uploading policies, the user is expected to provide the paths to a
Terraform JSON file and to a policy JSON file.

Resolves: INFL-300