Bugfix: various bugs and fixes
This commit includes various bugs and fixes to the TUI and CLI. Most
prominently, the code used to create the asset list is extracted to a
`ChooserViewport` model which can be reused. Both the environment
chooser and the drift chooser use this model now.

The display of the codify method output in the TUI is also fixed.