Add task utility and version number
This commit introduces a shell script called `task` to the root
directory. This script can be used during development and CI/CD
processes to perform common tasks such as compiling the CLI, running
unit tests and more.

When compiling the CLI, the script also generates a version number from
the latest commit hash and the current date. The CLI is also extended
with a `version` command that prints the version string and exits.

Relates to: INFL-322