Force configuring a profile when doesn't exist
This commit modifies the behavior of the CLI (and TUI) when a
configuration profile does not exist. Previously, if a profile did not
exist and a key-pair was not provided via command line flags, the CLI
would ask for a key-pair and continue. With this change, the user will
be forwarded to the configuration process, so that a profile is created
and properly stored. After configuration is done, the CLI will continue
with the actual command executed by the user.

Resolves: INFL-322