Add ability to run a classification rule
The classification client is updated to allow running a classification
rule against an account's inventory. This currently is only available
via the /exclusions/test API, and requires first loading the rule, which
itself is only possible by listing all of the rules.

A direct API should be available in the near future.

The list command is modified to accept input. The only input flag
currently supported is whether to base64-decode rules Rego code or not.
The default is to decode them, which was the behavior before this flag
was introduced.

Resolves: INFL-674