Change branding from Infralight to Firefly
As the company decided to change its name from "Infralight" to
"Firefly", the braning of the CLI codebase is hereby modified to reflect
that decision.

All relevant references to the previous name are changed appropriately.
The name of the product is now also a constant in the version package
(`version.Product`), and all references use the constant instead of
having the name hardcoded.

The CLI will also search for configuration profiles under the "firefly"
name, and fallback to the previous name if not found. This makes
upgrading the CLI version easier for customers already using it.

The only locations where the original name still exist are:

- The project URL (and thus package import statements).
- The API URL.
- The Logo in the README file.

Resolves: INFL-877